LPS-123796 Move components out of PageRenderer
